[style] clean up some warnings

+++ src/common/fs_util.h
@@ -20,8 +20,8 @@
  * Author: charisu
  */
 
-#ifndef  SRC_COMMON_FS_UTIL_H_
-#define  SRC_COMMON_FS_UTIL_H_
+#ifndef SRC_COMMON_FS_UTIL_H_
+#define SRC_COMMON_FS_UTIL_H_
 
 #include <glog/logging.h>
 #include <string>
@@ -32,8 +32,8 @@ namespace curve {
 namespace common {
 
 // 计算path2相对于path1的相对路径
-inline std::string CalcRelativePath(const std::string& path1,
-                                    const std::string& path2) {
+inline std::string CalcRelativePath(const std::string &path1,
+                                    const std::string &path2) {
     if (path1.empty() || path2.empty()) {
         return "";
     }
@@ -42,23 +42,24 @@ inline std::string CalcRelativePath(const std::string& path1,
     SplitString(path1, "/", &dirs1);
     SplitString(path2, "/", &dirs2);
     int unmatchedIndex = 0;
-    while (unmatchedIndex < dirs1.size() && unmatchedIndex < dirs2.size()) {
+    while (unmatchedIndex < static_cast<int>(dirs1.size()) &&
+           unmatchedIndex < static_cast<int>(dirs2.size())) {
         if (dirs1[unmatchedIndex] != dirs2[unmatchedIndex]) {
             break;
         }
         unmatchedIndex++;
     }
     std::string rpath;
-    if (unmatchedIndex == dirs1.size()) {
+    if (unmatchedIndex == static_cast<int>(dirs1.size())) {
         rpath.append(".");
     }
-    for (int i = 0; i < dirs1.size() - unmatchedIndex; ++i) {
+    for (int i = 0; i < static_cast<int>(dirs1.size()) - unmatchedIndex; ++i) {
         if (i > 0) {
             rpath.append("/");
         }
         rpath.append("..");
     }
-    for (int i = unmatchedIndex; i < dirs2.size(); ++i) {
+    for (int i = unmatchedIndex; i < static_cast<int>(dirs2.size()); ++i) {
         rpath.append("/");
         rpath.append(dirs2[i]);
     }
@@ -66,8 +67,7 @@ inline std::string CalcRelativePath(const std::string& path1,
 }
 
 // Check whether the path2 is the subpath of path1
-inline bool IsSubPath(const std::string& path1,
-                      const std::string& path2) {
+inline bool IsSubPath(const std::string &path1, const std::string &path2) {
     return StringStartWith(CalcRelativePath(path1, path2), "./");
 }
 
@@ -75,4 +75,3 @@ inline bool IsSubPath(const std::string& path1,
 }  // namespace curve
 
 #endif  // SRC_COMMON_FS_UTIL_H_
-
